pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chddecg"
version = "0.1.0"
description = "Production-oriented refactor of the CHDdECG notebook pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24,<3.0",
  "pandas>=2.1,<3.0",
  "scipy>=1.11,<2.0",
  "scikit-learn>=1.4,<2.0",
  "matplotlib>=3.8,<4.0",
  "seaborn>=0.13,<1.0",
  "tqdm>=4.66,<5.0",
  "PyWavelets>=1.5,<2.0",
  "wfdb>=4.1,<5.0",
  "imbalanced-learn>=0.12,<1.0",
  "PyYAML>=6.0,<7.0",
  "tensorflow>=2.16,<2.18"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
